refactor(backend): extract RAID routes into dedicated system_raid module
